+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Post v1 a change is further considered breaking if the state file format is inco
 
 * Added private message to players when they die. ([#188](https://github.com/hjwylde/werewolf/issues/188))
 * Added the Medusa role. ([#99](https://github.com/hjwylde/werewolf/issues/99))
+* Added `unvote` command. ([#213](https://github.com/hjwylde/werewolf/issues/213))
 
 *Revisions*